feat(examples): add documentation and walkthroughs for examples
This commit adds documentation along with walkthroughs for how to use
`jco` when building and deploying WebAssembly components.
