+++ examples/README.md
@@ -1,0 +1,16 @@
+# Examples
+
+This folder contains examples of using `jco`, `componentize-js`, and related Javascript WebAssembly runtime tooling.
+
+The following sections are present:
+
+- [Example Components](./components) - "Just show me the code" examples of how to build JS projects that use `jco`
+- [Guides](./guides) - Walkthroughs of concepts and functionality to keep in mind while using `jco` and associated tooling
+
+> [!NOTE]
+> Note sure what WebAssembly "Components" are, relative to WebAssembly modules?
+>
+> For an in-depth multi-language guide to the concepts and advanced features behind the Component Model,
+> [read the Component Model book][cm-book].
+
+[cm-book]: https://component-model.bytecodealliance.org/
