+++ examples/components/README.md
@@ -1,0 +1,14 @@
+## Example Components
+
+Most (if not all) individual example projects are standard Javascript projects commonplace ecosystem tooling, whether
+for server side ([NodeJS][nodejs] -- `npm`, etc) or the browser.
+
+A brief description of the examples contained in this folder:
+
+| Example                                          | Description                                                                                      |
+|--------------------------------------------------|--------------------------------------------------------------------------------------------------|
+| [`add`](./add)                                   | `export`s basic functionality with simple types                                                  |
+| [`string-reverse`](./string-reverse)             | `export`s basic functionality with a slightly more involved WIT interface and more complex types |
+| [`string-reverse-upper`](./string-reverse-upper) | `import`s functionality to build more advanced computation to `export`                           |
+
+[nodejs]: https://nodejs.org
