+++ examples/components/add/add.js
@@ -1,0 +1,7 @@
+/**
+ * This Javascript will be interpreted by `jco` and turned into a
+ * WebAssembly binary with a single export (this `add` function).
+ */
+export function add(x, y) {
+  return x + y;
+}
