--- examples/components/add/add.mjs
@@ -1,7 +1,0 @@
-/**
- * This Javascript will be interpreted by `jco` and turned into a
- * WebAssembly binary with a single export (this `add` function).
- */
-export function add(x, y) {
-  return x + y;
-}
