+++ examples/components/add/run-transpiled.js
@@ -1,4 +1,4 @@
 // If this import listed below is missing, please run `npm run transpile`
-import { add } from "./dist/transpiled/add.mjs";
+import { add } from "./dist/transpiled/add.js";
 
 console.log("1 + 2 = " + add(1, 2));
